Refactored and renamed imageRecords for clarity
