+++ app/direct/[username]/page.tsx
@@ -1,7 +1,7 @@
 import { BeastCard } from "@/app/components/gallery/BeastCard";
 import { Card } from "@/app/components/ui/card";
 import Link from "next/link";
-import { getImageForUser } from "@/lib/data";
+import { fetchUserImageRecord } from "@/lib/imageRecords";
 
 // Applying 'any' workaround for props due to Next.js 15 build issue
 export default async function DirectPage(props: any) {
@@ -11,7 +11,7 @@ export default async function DirectPage(props: any) {
     return <div>Error: Username not found in parameters.</div>;
   }
 
-  const image = await getImageForUser(username);
+  const image = await fetchUserImageRecord(username);
   
   return (
     <div className="w-full flex flex-col items-center pt-8">
